chore: move inputs to top level
because the constructor is its own go module at the root, we cannot put the input go module in a separate root path
