+++ Taskfile.yml
@@ -13,10 +13,10 @@ includes:
     optional: true
     taskfile: ./bindings/go/constructor/Taskfile.yml
     dir: ./bindings/go/constructor
-  bindings/go/constructor/input/file:
+  bindings/go/input/file:
     optional: true
-    taskfile: ./bindings/go/constructor/input/file/Taskfile.yml
-    dir: ./bindings/go/constructor/input/file
+    taskfile: ./bindings/go/input/file/Taskfile.yml
+    dir: ./bindings/go/input/file
   bindings/go/descriptor/v2:
     optional: true
     taskfile: ./bindings/go/descriptor/v2/Taskfile.yml
