+++ bindings/go/input/file/doc.go
