+++ bindings/go/input/file/go.mod
@@ -1,4 +1,4 @@
-module ocm.software/open-component-model/bindings/go/constructor/input/file
+module ocm.software/open-component-model/bindings/go/input/file
 
 go 1.24.2
 
@@ -6,7 +6,7 @@ require (
 	github.com/gabriel-vasile/mimetype v1.4.9
 	github.com/stretchr/testify v1.10.0
 	ocm.software/open-component-model/bindings/go/blob v0.0.3
-	ocm.software/open-component-model/bindings/go/constructor v0.0.0-20250626125027-c7f0912ab5fd
+	ocm.software/open-component-model/bindings/go/constructor v0.0.0-20250701105956-02bd6cbc8499
 	ocm.software/open-component-model/bindings/go/runtime v0.0.2
 )
 
@@ -32,7 +32,7 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	ocm.software/open-component-model/bindings/go/descriptor/runtime v0.0.0-20250625081434-d598b670071d // indirect
 	ocm.software/open-component-model/bindings/go/descriptor/v2 v2.0.0-alpha2 // indirect
-	ocm.software/open-component-model/bindings/go/oci v0.0.0-20250623110822-9dca0bcd1765 // indirect
+	ocm.software/open-component-model/bindings/go/oci v0.0.0-20250626125027-c7f0912ab5fd // indirect
 	oras.land/oras-go/v2 v2.6.0 // indirect
 	sigs.k8s.io/yaml v1.5.0 // indirect
 )
