+++ bindings/go/input/file/spec/v1/spec.go
