+++ bindings/go/input/file/spec/v1/zz_generated.deepcopy.go
