+++ bindings/go/input/file/spec/v1/zz_generated.ocm_type.go
